Update server to create users via pathom not ring
Previously we would create users on the ring handler when they didn't exist.

These changes instead move the user creation to a pathom plugin. This keeps all of
our data and mutations in pathom's realm.

We also update the broadcaster to pass up a query instead of a component. The query
is grabbed from the cljc component, but we also could push any query we want up.
This has the nice benefit of being able to update our clients with whatever data we
want.