Push the wormhole connection logic into sente
Sente provides the exact abstraction we want:
- Each room corresponds to a sente user
This means that as long as there is a browser tab with a room open, we won't break the
wormhole connection. When the last browser tab for a room is open, close the open connection
if there is one

There are some drawbacks here:
- logic is outside of fulcro/pathom (not consolidated)
- could be expensive if the `user-id-fn` grows too large
- broadcasting logic gets more complicated and error prone

Still need to figure out a faster way to break the connection, sente's user disconnect fires
after about six seconds.

Overall I am happy with this approach though.